requirements.txt
tweepy==3.9.0
mysql-connector==2.2.9
mysqlclient==1.4.6
oauthlib==3.1.0
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
json5==0.9.1
jsonschema==3.2.0
